# perf: remove useless copies in e4 mul by elem