--- a/field/babybear/extensions/e4.go
+++ b/field/babybear/extensions/e4.go
@@ -78,10 +78,8 @@ func (z *E4) SetOne() *E4 {
 
 // MulByElement multiplies an element in E4 by an element in fr
 func (z *E4) MulByElement(x *E4, y *fr.Element) *E4 {
-	var yCopy fr.Element
-	yCopy.Set(y)
-	z.B0.MulByElement(&x.B0, &yCopy)
-	z.B1.MulByElement(&x.B1, &yCopy)
+	z.B0.MulByElement(&x.B0, y)
+	z.B1.MulByElement(&x.B1, y)
 	return z
 }