--- a/field/generator/internal/templates/extensions/e4_test.go.tmpl
+++ b/field/generator/internal/templates/extensions/e4_test.go.tmpl
@@ -3,6 +3,8 @@ import (
 
 	"github.com/leanovate/gopter"
 	"github.com/leanovate/gopter/prop"
+
+	fr "{{ .FieldPackagePath }}"
 )
 
 // ------------------------------------------------------------
@@ -249,6 +251,17 @@ func BenchmarkE4Mul(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		a.Mul(&a, &c)
+	}
+}
+
+func BenchmarkE4MulByElement(b *testing.B) {
+	var a E4
+	var c fr.Element
+	_, _ = a.SetRandom()
+	_, _ = c.SetRandom()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		a.MulByElement(&a, &c)
 	}
 }