# Description

At least for small fields, the copy is useless / hurts perf. Likely same for other towers?

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkE4MulByElement-16     6.74          5.95          -11.79%
BenchmarkE4MulByElement-16     6.76          5.86          -13.29%
```